feat(nuxt): export event handler
