+++ .changeset/green-donkeys-decide.md
@@ -1,0 +1,5 @@
+---
+"@logto/nuxt": minor
+---
+
+export logtoEventHandler
